+++ src/util/on-create.mjs
@@ -53,3 +53,42 @@ export const onCreate = (
     };
     Hooks.on('preCreateItem', handleOnCreate);
 }
+
+/**
+ * @param {ItemPF} item
+ * @param {string} key
+ * @param {string} compendiumId
+ * @param {((name: string) => boolean) | string | boolean} nameFunc
+ * @param {boolean} isEditable
+ * @param {() => object} [flagValueFunc]
+ *
+ * @returns {boolean}
+ */
+export const onRenderCreate = (item, key, compendiumId, nameFunc, isEditable, flagValueFunc) => {
+    const hasFlag = item.hasItemBooleanFlag(key);
+    if (!hasFlag) {
+        const name = item?.name?.toLowerCase() ?? "";
+        const hasCompendiumId = itemHasCompendiumId(item, compendiumId);
+
+        const nameMatch = typeof nameFunc === 'string'
+            ? nameFunc === name
+            : typeof nameFunc === 'boolean'
+                ? nameFunc
+                : nameFunc(name);
+        if (isEditable && (nameMatch || hasCompendiumId)) {
+            item.addItemBooleanFlag(key);
+            if (flagValueFunc) {
+                try {
+                    const value = flagValueFunc()
+                    item.setFlag(MODULE_NAME, key, value);
+                    return true;
+                }
+                catch (e) {
+                    console.error("Shouldn't happen", e);
+                }
+            }
+        }
+        return false;
+    }
+    return true;
+}